Former-commit-id: 4fad21178635ddab8f6ce5ccbfee03f2de6c333b
Former-commit-id: 5fac25b1878fdec4573aa6021977286ff1384fe7

+++ registration-processor/init/registration-processor-packet-receiver-stage/src/main/java/io/mosip/registration/processor/packet/receiver/stage/PacketReceiverStage.java
@@ -132,7 +132,15 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 	 * @param routingContext the routing context
 	 */
 	public void failure(RoutingContext routingContext) {
-		this.setResponseWithDigitalSignature(routingContext, globalExceptionHandler.handler(routingContext.failure()), APPLICATION_JSON);
+		try {
+			deleteFile(getFileFromCtx(routingContext));
+		} catch (IOException e) {
+			regProcLogger.error(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
+					"", e.getMessage() + ExceptionUtils.getStackTrace(e));
+		}
+		this.setResponseWithDigitalSignature(routingContext, globalExceptionHandler.handler(routingContext.failure()),
+				APPLICATION_JSON);
+
 	}
 
 
@@ -157,12 +165,8 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 					"", e.getMessage() + ExceptionUtils.getStackTrace(e));
 			throw new UnexpectedException(e.getMessage());
 		} finally {
-			if (file != null) {
-				if (file.exists()) {
-					deleteFile(file);
-				}
-			}
-		 }
+			deleteFile(file);
+		}
 		regProcLogger.debug(LoggerFileConstant.SESSIONID.toString(), LoggerFileConstant.REGISTRATIONID.toString(),
 				"", "PacketReceiverStage::processPacket()::exit");
 
@@ -205,9 +209,13 @@ public class PacketReceiverStage extends MosipVerticleAPIManager {
 	 * @param file
 	 *            the file
 	 */
-	private void deleteFile(File file) {
+	public void deleteFile(File file) {
 		try {
-			FileUtils.forceDelete(file);
+			if (file != null) {
+				if (file.exists()) {
+					FileUtils.forceDelete(file);
+				}
+			}
 		} catch (IOException e) {
 			throw new UnexpectedException(e.getMessage());
 		}
